requirements.txt
langchain==0.0.136
llama_index==0.5.12
openai==0.27.2
python-dotenv==1.0.0
